refactor(WebAPI): Moved AutoPauseActiveDownloads and RecoverInterruptedDownloads to separate command handlers
